+++ gplearn/tests/test_functions.py
@@ -4,14 +4,17 @@
 #
 # License: BSD 3 clause
 
+import pickle
+
 import numpy as np
 from numpy import maximum
-from sklearn.datasets import load_boston
+from sklearn.datasets import load_boston, load_breast_cancer
 from sklearn.utils.testing import assert_equal, assert_raises
 from sklearn.utils.validation import check_random_state
 
 from gplearn.functions import _protected_sqrt, make_function
-from gplearn.genetic import SymbolicTransformer
+from gplearn.genetic import SymbolicRegressor, SymbolicTransformer
+from gplearn.genetic import SymbolicClassifier
 
 # load the boston dataset and randomly permute it
 rng = check_random_state(0)
@@ -19,6 +22,12 @@ boston = load_boston()
 perm = rng.permutation(boston.target.size)
 boston.data = boston.data[perm]
 boston.target = boston.target[perm]
+
+# load the breast cancer dataset and randomly permute it
+cancer = load_breast_cancer()
+perm = check_random_state(0).permutation(cancer.target.size)
+cancer.data = cancer.data[perm]
+cancer.target = cancer.target[perm]
 
 
 def test_validate_function():
@@ -29,6 +38,8 @@ def test_validate_function():
     # non-integer arity
     assert_raises(ValueError, make_function, _protected_sqrt, 'sqrt', '1')
     assert_raises(ValueError, make_function, _protected_sqrt, 'sqrt', 1.0)
+    # non-bool wrap
+    assert_raises(ValueError, make_function, _protected_sqrt, 'sqrt', 1, 'f')
     # non-matching arity
     assert_raises(ValueError, make_function, _protected_sqrt, 'sqrt', 2)
     assert_raises(ValueError, make_function, maximum, 'max', 1)
@@ -79,3 +90,76 @@ def test_function_in_program():
     formula = est._programs[0][906].__str__()
     expected_formula = 'sub(logical(X6, add(X11, 0.898), X10, X2), X5)'
     assert_equal(expected_formula, formula, True)
+
+
+def test_parallel_custom_function():
+    """Regression test for running parallel training with custom functions"""
+
+    def _logical(x1, x2, x3, x4):
+        return np.where(x1 > x2, x3, x4)
+
+    logical = make_function(function=_logical,
+                            name='logical',
+                            arity=4)
+    est = SymbolicRegressor(generations=2,
+                            function_set=['add', 'sub', 'mul', 'div', logical],
+                            random_state=0,
+                            n_jobs=2)
+    est.fit(boston.data, boston.target)
+    _ = pickle.dumps(est)
+
+    # Unwrapped functions should fail
+    logical = make_function(function=_logical,
+                            name='logical',
+                            arity=4,
+                            wrap=False)
+    est = SymbolicRegressor(generations=2,
+                            function_set=['add', 'sub', 'mul', 'div', logical],
+                            random_state=0,
+                            n_jobs=2)
+    est.fit(boston.data, boston.target)
+    assert_raises(AttributeError, pickle.dumps, est)
+
+    # Single threaded will also fail in non-interactive sessions
+    est = SymbolicRegressor(generations=2,
+                            function_set=['add', 'sub', 'mul', 'div', logical],
+                            random_state=0)
+    est.fit(boston.data, boston.target)
+    assert_raises(AttributeError, pickle.dumps, est)
+
+
+def test_parallel_custom_transformer():
+    """Regression test for running parallel training with custom transformer"""
+
+    def _sigmoid(x1):
+        with np.errstate(over='ignore', under='ignore'):
+            return 1 / (1 + np.exp(-x1))
+
+    sigmoid = make_function(function=_sigmoid,
+                            name='sig',
+                            arity=1)
+    est = SymbolicClassifier(generations=2,
+                             transformer=sigmoid,
+                             random_state=0,
+                             n_jobs=2)
+    est.fit(cancer.data, cancer.target)
+    _ = pickle.dumps(est)
+
+    # Unwrapped functions should fail
+    sigmoid = make_function(function=_sigmoid,
+                            name='sig',
+                            arity=1,
+                            wrap=False)
+    est = SymbolicClassifier(generations=2,
+                             transformer=sigmoid,
+                             random_state=0,
+                             n_jobs=2)
+    est.fit(cancer.data, cancer.target)
+    assert_raises(AttributeError, pickle.dumps, est)
+
+    # Single threaded will also fail in non-interactive sessions
+    est = SymbolicClassifier(generations=2,
+                             transformer=sigmoid,
+                             random_state=0)
+    est.fit(cancer.data, cancer.target)
+    assert_raises(AttributeError, pickle.dumps, est)