+++ doc/changelog.rst
@@ -9,6 +9,14 @@ Version 0.5.0
 
 - Nothing yet!
 
+Version 0.4.1 - 1 Jun 2019
+---------------------------
+
+- Fixed a bug with multi-processing and custom functions, allowing pickling of
+  models with custom functions, fitness metrics or classifier transformaers.
+  ``joblib`` 0.13.0 or newer required in order to take advantage of this
+  release in order to wrap functions for pickling saved models.
+
 Version 0.4.0 - 23 Apr 2019
 ---------------------------